net: lwm2m: move non-public data out of shared TLV include
